Fixed hiearchy issue with Resource & reworded naming.
It's better to use A::B::C than make assumptions about A and B when
declaring C.  For example, Sequencescape::Api might not always be a
class!

The associations were broken because of the call order of initialize.  I
still don't know if it's better to declare all methods inside modules
and include/extend them.