Allow applications to specify namespaces for models.
It's become clear that the Sequencescape namespace are the root of the
object hierarchy.  Each application will then, more than likely, extend
these base classes to add behaviour.  To do that, they cannot simply use
the Sequencescape namespace, so passing a module/class in :namespace to
the Sequencescape::Api constructor will allow you to override this
behaviour.