Started working on better attribute interface.
This uses ActiveModel dirty to ensure that only changed attributes are
sent to the server.  It is specifically coded to deal with the batches
in our pipelines and isn't guaranteed for anything else.  It requires
all attributes to be declared up front, which may or may not be a
problem depending on the migration path of the API.