Validations must include associations.
When validating an object we have to check any associated objects (don't
worry, no cycles in this object graph!) and also maintain all of the
errors correct.  There is one case where the validations should not be
run and that is when using the UUID of an object on an association, in
which case we're assuming the object we reference is valid.