Implemented all of the required attributes on the models.
All of the attributes for the models should be present but if they are
not the code will not fall over as I've made it optionally assign values
(for instance, request uses asset but different types of assets have
different types of attributes).  Did some work on composed_of too so
that the composed objects behave a little cleaner.