Use request/response contracts to test behaviour.
Completely removed all of the specs that existed before and started
covering the code with specs that use contracts: the idea that a request
illicits a response.  Whilst this has reduced the number of tests (and
probably coverage for the moment) it has meant that mocks aren't being
used and so less churn should be seen in the specs themselves.

Discovered several issues within the code in doing this:

* Fixed an issue where belongs_to would not load.
* Fixed an issue where class names weren't properly resolved.
* Fixed an issue where updates weren't sending deltas.
* Fixed an issue where create was sending blank.
* Fixed an issue with retrieving all instances of a model.

All of which were pretty fundamental!