Fix issue with write-only or read-only attributes.
Unfortunately the read-only fix broke the changes code and the
write-only attributes were completely broken.  ActiveModel requires a
reader method to read the changes, and we need to be able to signal that
a value has changed.  So this code always generates both a reader and a
writer but hides the ones that are not needed.