Enable grouping of attributes.
Samples have a highly structured JSON from the API and this commit
provides a convenience to that structure so that the code can use dotted
notation (sample.sanger.name) rather than hash based versions
(sample.sanger['name']).

This has only been checked for read-only abilities, which is what is
needed to support the pulldown pipeline.