[#30992811] Pool information for a plate is structured
Added the ability to cache responses from the server, and to force
reloading if necessary.  Especially useful for the wells, and that's the
only place it's used at the moment.

Enabled an after_load callback for all resources, which allows the plate
to update the pool information.

Added Plate#populate_wells_with_pool method so that the pulldown
application can selective fill this information into the wells.