Proxy defers to object for uuid_only?
This change is in response to changes in Rails 4
which result in additional responds_to? calls
on the proxy object which are not correctly handled
for belongs_to associations established with a uuid.
The api attempts to lookup the object with a non-
existent read action url.

It may prove more robust in the long term to adjust
the attempt to load the object. But this is likely
to have wider impact elsewhere.